Moved functions out of spec_class
